Create fullScreenVertexShaderOptions utils
